lucet-idl: move C generator to be all one file
